requirements.txt
# Core dependencies
python-dotenv>=1.0.0
Pillow>=10.0.0

# Model providers
google-generativeai>=0.3.0
anthropic>=0.7.0
openai>=1.0.0

# Image processing
numpy>=1.24.0

# Type hints
typing-extensions>=4.8.0

# Development tools
pytest>=7.4.0
black>=23.9.0
isort>=5.12.0
flake8>=6.1.0
mypy>=1.5.0
